Fix diagnostics imports and formatting
