+++ build_and_wrap.py
@@ -4,6 +4,7 @@
 from __future__ import annotations
 
 import json
+import logging
 import os
 from pathlib import Path
 from typing import Any, Callable
@@ -16,6 +17,8 @@ from monGARS.mlops.artifacts import (
     write_wrapper_bundle,
 )
 from monGARS.mlops.dataset import prepare_instruction_dataset
+from monGARS.mlops.diagnostics.analysis import analyse_cuda_state
+from monGARS.mlops.diagnostics.cuda_metrics import gather_cuda_metrics
 from monGARS.mlops.exporters import export_gguf, merge_lora_adapters
 from monGARS.mlops.model import load_4bit_causal_lm, summarise_device_map
 from monGARS.mlops.training import (
@@ -32,6 +35,13 @@ from monGARS.mlops.utils import (
     ensure_dependencies,
     ensure_directory,
 )
+
+logger = logging.getLogger(__name__)
+
+try:  # pragma: no cover - optional dependency
+    import torch  # type: ignore[import-not-found]
+except ModuleNotFoundError:  # pragma: no cover - optional dependency
+    torch = None  # type: ignore[assignment]
 
 MODEL_ID = os.environ.get("MODEL_ID", "cognitivecomputations/Dolphin3.0-Llama3.1-8B")
 DATASET_NAME = os.environ.get("DATASET", "yahma/alpaca-cleaned")
@@ -159,6 +169,49 @@ def _maybe_update_registry(registry_path: str | None, summary: dict[str, Any]) -
         )
 
 
+def _log_oom_report(analysis: dict[str, Any]) -> None:
+    status = analysis.get("status", "unknown")
+    logger.info("OOM risk classification: %s", status)
+
+    for device in analysis.get("devices", []) or []:
+        index = device.get("index", "?")
+        free = device.get("free_gib")
+        ratio = device.get("free_ratio")
+        device_status = device.get("status", "unknown")
+        if free is not None and ratio is not None:
+            logger.info(
+                "Device %s: %.2f GiB free (%.1f%%) -> %s",
+                index,
+                free,
+                ratio * 100,
+                device_status,
+            )
+        else:
+            logger.info("Device %s: status %s", index, device_status)
+        for recommendation in device.get("recommendations", []):
+            logger.info("Device %s recommendation: %s", index, recommendation)
+
+
+def _raise_on_critical(analysis: dict[str, Any], fail_on_critical: bool) -> None:
+    if not fail_on_critical or analysis.get("status") != "critical":
+        return
+
+    recommendations: list[str] = [
+        recommendation
+        for device in analysis.get("devices", []) or []
+        for recommendation in device.get("recommendations", [])
+    ]
+    guidance = "\n".join(f"  - {text}" for text in recommendations)
+    if not guidance:
+        guidance = "  - See diagnose_unsloth guidance."
+
+    raise RuntimeError(
+        "Insufficient CUDA headroom for QLoRA fine-tuning.\n"
+        "VRAM diagnostics flagged a critical OOM risk.\n"
+        f"Recommended mitigations:\n{guidance}"
+    )
+
+
 def evaluate_oom_headroom(
     *,
     min_free_gib: float = OOM_MIN_FREE_GIB,
@@ -166,21 +219,14 @@ def evaluate_oom_headroom(
     fail_on_critical: bool = FAIL_ON_CRITICAL_OOM,
     torch_module: Any | None = None,
     gather_metrics: Callable[[Any], dict[str, Any] | None] | None = None,
-    analyse_cuda_state: Callable[..., dict[str, Any]] | None = None,
+    analyse_state: Callable[..., dict[str, Any]] | None = None,
 ) -> dict[str, Any]:
     """Assess CUDA headroom and optionally abort if the risk is critical."""
 
     skip_reason: str | None = None
     cuda_payload: dict[str, Any] | None = None
 
-    if analyse_cuda_state is None:
-        from scripts.diagnose_unsloth import _analyse_cuda_state as analyse_cuda_state  # type: ignore
-
-    if torch_module is None:
-        try:
-            import torch as torch_module  # type: ignore
-        except ModuleNotFoundError:
-            torch_module = None
+    torch_module = torch_module if torch_module is not None else torch
 
     if torch_module is None:
         skip_reason = "torch_missing"
@@ -191,57 +237,25 @@ def evaluate_oom_headroom(
     elif not torch_module.cuda.is_available():
         skip_reason = "cuda_unavailable"
     else:
-        if gather_metrics is None:
-            from scripts.diagnose_unsloth import _gather_cuda_metrics as _gather_cuda_metrics  # type: ignore
-
-            def gather_metrics(module: Any) -> dict[str, Any] | None:  # type: ignore
-                return _gather_cuda_metrics(
-                    module,
-                    lambda: list(range(module.cuda.device_count())),
-                )
-
-        cuda_payload = gather_metrics(torch_module)
+        metrics_fn = gather_metrics
+        if metrics_fn is None:
+            metrics_fn = lambda module: gather_cuda_metrics(  # type: ignore[assignment]
+                module, lambda: list(range(module.cuda.device_count()))
+            )
+        cuda_payload = metrics_fn(torch_module)
         if cuda_payload is None:
             skip_reason = "cuda_metrics_unavailable"
 
-    analysis = analyse_cuda_state(
+    analyser = analyse_state or analyse_cuda_state
+    analysis = analyser(
         cuda_payload,
         min_free_gib=min_free_gib,
         min_free_ratio=min_free_ratio,
         skip_reason=skip_reason,
     )
 
-    status = analysis.get("status", "unknown")
-    print(f"OOM risk classification: {status}")
-
-    devices = analysis.get("devices") or []
-    for device in devices:
-        index = device.get("index", "?")
-        free = device.get("free_gib")
-        ratio = device.get("free_ratio")
-        device_status = device.get("status", "unknown")
-        if free is not None and ratio is not None:
-            print(
-                f"  - Device {index}: {free:.2f} GiB free ({ratio:.1%}) -> {device_status}"
-            )
-        else:
-            print(f"  - Device {index}: status {device_status}")
-        for recommendation in device.get("recommendations", []):
-            print(f"      recommendation: {recommendation}")
-
-    if fail_on_critical and status == "critical":
-        recommendations: list[str] = []
-        for device in devices:
-            recommendations.extend(device.get("recommendations", []))
-        guidance = (
-            "\n".join(f"  - {text}" for text in recommendations)
-            or "  - See diagnose_unsloth guidance."
-        )
-        raise RuntimeError(
-            "Insufficient CUDA headroom for QLoRA fine-tuning.\n"
-            "VRAM diagnostics flagged a critical OOM risk.\n"
-            f"Recommended mitigations:\n{guidance}"
-        )
+    _log_oom_report(analysis)
+    _raise_on_critical(analysis, fail_on_critical)
 
     return analysis
