+++ monGARS/mlops/diagnostics/__init__.py
@@ -1,0 +1,17 @@
+"""Diagnostics helpers for CUDA headroom and Unsloth integration."""
+
+from __future__ import annotations
+
+from .analysis import analyse_cuda_state
+from .cli import main as diagnostics_main
+from .cuda_metrics import gather_cuda_metrics
+from .environment import configure_logging, gather_environment, import_optional
+
+__all__ = [
+    "analyse_cuda_state",
+    "configure_logging",
+    "diagnostics_main",
+    "gather_cuda_metrics",
+    "gather_environment",
+    "import_optional",
+]
